feat: group table properties into 3 sections
