+++ frontend/src/apps/slides/components/TableSection.vue
@@ -32,44 +32,6 @@
 		>
 			<Switch :modelValue="headers.column" @update:modelValue="toggleHeaderColumn" />
 		</PropertyRow>
-		<PropertyRow
-			label="Banded Rows"
-			class="cursor-pointer"
-			@click="toggleFromRow($event, () => setBandedRows(!activeElement.bandedRows))"
-		>
-			<Switch :modelValue="activeElement.bandedRows || false" @update:modelValue="setBandedRows" />
-		</PropertyRow>
-		<PropertyRow v-if="activeElement.bandedRows" label="Band Color">
-			<ColorPicker
-				:modelValue="activeElement.bandColor || getDefaultBandColor(activeElement.color)"
-				@update:modelValue="bandColor.set"
-				@colordown="bandColor.begin"
-				@colorup="bandColor.commit"
-			/>
-		</PropertyRow>
-		<PropertyRow label="Cell Fill">
-			<ColorPicker :modelValue="editorStyles.cellFill" @update:modelValue="setCellFill" />
-		</PropertyRow>
-		<PropertyRow label="Grid Color">
-			<ColorPicker
-				:modelValue="activeElement.gridColor || getDefaultGridColor(activeElement.color)"
-				@update:modelValue="gridColor.set"
-				@colordown="gridColor.begin"
-				@colorup="gridColor.commit"
-			/>
-		</PropertyRow>
-		<NumberControl
-			:modelValue="activeElement.gridWidth ?? 1"
-			label="Grid Weight"
-			suffix="px"
-			:min="0"
-			:max="3"
-			:max-digits="3"
-			:step="0.5"
-			@update:modelValue="gridWidth.set"
-			@change-start="gridWidth.begin"
-			@change-end="gridWidth.commit"
-		/>
 	</Section>
 </template>
 
@@ -78,16 +40,11 @@ import { computed } from 'vue'
 
 import { Switch } from 'frappe-ui'
 
-import ColorPicker from '@/apps/slides/components/controls/ColorPicker.vue'
 import NumberControl from '@/apps/slides/components/controls/NumberControl.vue'
 import PropertyRow from '@/apps/slides/components/controls/PropertyRow.vue'
 import Section from '@/apps/slides/components/controls/Section.vue'
 
 import { activeElement } from '@/apps/slides/stores/element'
-import { useTextEditor } from '@/apps/slides/composables/useTextEditor'
-import { setCellFill } from '@/apps/slides/utils/tableCells'
-import { setElementProperty, useElementProperty } from '@/apps/slides/composables/editProperty'
-import { getDefaultBandColor, getDefaultGridColor } from '@/apps/slides/utils/color'
 import { getTableSize, getTableHeaders } from '@/apps/slides/utils/tableWidths'
 import {
 	setRowCount,
@@ -95,8 +52,6 @@ import {
 	toggleHeaderRow,
 	toggleHeaderColumn,
 } from '@/apps/slides/utils/tableStructure'
-
-const { editorStyles } = useTextEditor()
 
 const tableSize = computed(() => getTableSize(activeElement.value.content))
 
@@ -106,12 +61,4 @@ const headers = computed(() => getTableHeaders(activeElement.value.content))
 const toggleFromRow = (e, toggle) => {
 	if (!e.target.closest('button')) toggle()
 }
-
-const setBandedRows = (value) => setElementProperty('bandedRows', value)
-
-const bandColor = useElementProperty('bandColor')
-
-const gridColor = useElementProperty('gridColor')
-
-const gridWidth = useElementProperty('gridWidth')
 </script>
